requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
boto3>=1.34.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0